requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
sentence-transformers>=2.2
torch>=2.0
requests>=2.31
tqdm>=4.66
pyyaml>=6.0
openpyxl>=3.1